refactor:#6 가독성 향상을 위한 컴포넌트 분리
