+++ src/components/features/myPlanPage/ContentsContainer.jsx
@@ -1,0 +1,5 @@
+const ContentsContainer = ({ children }) => {
+  return <div className="m-4 flex flex-col gap-2">{children}</div>;
+};
+
+export default ContentsContainer;
